requirements.txt
boto3==1.9.25
botocore==1.12.25
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
idna==2.7
jmespath==0.9.3
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
